+++ tests/testthat/mod/path/to/__module__.R
@@ -1,0 +1,9 @@
+#' @export
+b_fun_a <- function() {
+  "exported a"
+}
+
+#' @export
+`02_fun` <- function() {
+  "exported a"
+}